% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.factor.models.R
\encoding{UTF-8}
\name{fit.attribution}
\alias{fit.attribution}
\alias{fit.fundamental}
\title{fit.attribution}
\usage{
fit.attribution(fitdata, date.var, id.var, return.var, exposure.vars,
  weight.var = NULL, rob.stats = FALSE, full.resid.cov = FALSE,
  z.score = FALSE, stdReturn = FALSE, resid.EWMA = FALSE,
  lambda = 0.9, ...)

fit.fundamental(fitdata, date.var, id.var, return.var, exposure.vars,
  weight.var = NULL, rob.stats = FALSE, z.score = FALSE,
  stdReturn = TRUE, calc.inv = TRUE, cov.wgt = FALSE,
  parkinson = FALSE)
}
\arguments{
\item{date.var}{string name of the date column}

\item{id.var}{string name of the id column}

\item{return.var}{string name of the return column}

\item{weight.var}{string name of the weight column}

\item{rob.stats}{logical}

\item{full.resid.cov}{logical}

\item{z.score}{logical}

\item{resid.EWMA}{logical}

\item{calc.inv}{logical}

\item{cov.wgt}{logical}

\item{parkinson}{logical}

\item{exposures.var}{vector of string name(s) of the column(s) to use as exposures (attributes)}

\item{strReturns}{logical}

\item{date.var}{string name of the date column}

\item{id.var}{string name of the id column}

\item{return.var}{string name of the return column}

\item{weight.var}{string name of the weight column (used in regressions to estimate factor returns).}

\item{exposures.var}{vector of string name(s) of the column(s) to use as exposures (attributes)}

\item{rob.stats}{logical}

\item{z.score}{logical}

\item{strReturns}{logical}
}
\value{
data.table

list of risk model outputs where
*specific.risk* is the vector idiosyncratic volatility (or specific risk),
*factor.returns* is the matrix of factor returns estimated by the risk model,
*factor.loadings* is the matrix of factor loadings,
*factor.cov* is the matrix of factor covariances,
*cov.mat** is the matrix of variance/covariances,
*inverse.cov* is the inverse of the variance/covariance matrix, and
*fitdata* is the matrix of input data.
}
\description{
Function to calculate factor returns for portfolio attribution analysis.  The code has been adapted from the factorAnalyticsAddons package (Uthaisaad, 2017).

Function to calculate fundamental risk model. The code has been adapted from Kakushadze and Yu (2016).
}
\examples{
load("data/Stock.df.Rdata")
fit <- fit.attribution(fitdata = stock, date.var = 'DATE', id.var = 'TICKER', return.var = 'RETURN',
  weight.var = 'LOG.MARKETCAP', exposure.vars = c('NET.SALES','BOOK2MARKET','GICS.SECTOR'),
  rob.stats = TRUE, full.resid.cov = FALSE, z.score = FALSE,
  stdReturn = TRUE, resid.EWMA = TRUE)
load("data/Stock.df.Rdata")
fit <- fit.attribution(fitdata = stock, date.var = 'DATE', id.var = 'TICKER', return.var = 'RETURN',
  weight.var = 'LOG.MARKETCAP', exposure.vars = c('NET.SALES','BOOK2MARKET','GICS.SECTOR'),
  rob.stats = TRUE, full.resid.cov = FALSE, z.score = FALSE,
  stdReturn = TRUE, resid.EWMA = TRUE)
}
\references{
\url{https://github.com/chindhanai/factorAnalyticsAddons}

Kakushadze, Zura and Yu, Willie, Multifactor Risk Models and Heterotic CAPM (January 24, 2016). The Journal of Investment Strategies 5(4) (2016) 1-49. Available at SSRN: \url{https://ssrn.com/abstract=2722093}
}
\author{
Roger J. Bos, \email{roger.bos@gmail.com}

Roger J. Bos, \email{roger.bos@gmail.com}
}
\concept{factor models for attribution and risk}
