% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.factor.models.R
\encoding{UTF-8}
\name{fit.balanced.panel}
\alias{fit.balanced.panel}
\title{fit.balanced.panel}
\usage{
fit.balanced.panel(.data, .ids, .dates)
}
\arguments{
\item{.ids}{name of the identifier field that will be used for row names}

\item{.dates}{name of the data.table of rawdata}

\item{.dates}{name of the date field that will be used as column names}
}
\value{
data.table with equal number of observations in each month
}
\description{
Function to create a balanced data set (same number of observatons in each month)
}
\examples{
balanced <- fit.balanced.panel(rawdata, "symbol", "datadate")
}
\author{
Roger J. Bos, \email{roger.bos@gmail.com}
}
\concept{factor models for attribution and risk}
