% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.factor.models.R
\name{fit.contribution}
\alias{fit.contribution}
\title{fit.contribution}
\usage{
fit.contribution(fit, bm.wgt.var, port.wgt.var)
}
\arguments{
\item{fit}{list output from \code{"fit.attribution"}}

\item{bm.wgt.var}{string name of benchmark weight column (in fitdata passed to \code{"fit.attribution"})}

\item{port.wgt.var}{string name of portfolio weight column (in fitdata passed to \code{"fit.attribution"})}
}
\value{
list of periodic returns and summary returns
}
\description{
Helper function to show contributions nicely.  The periodic contributions from attribution cannot be cummulated to longer time periods so the carino adjustment is made as per Carino (1999).
}
\examples{
load("fit.factor.models/data/Stock.df.Rdata")
stock <- as.data.table(stock)
stock[TICKER \%in\% c('SUNW','ORCL','MSFT'), portfolioWeight := 1/3, by=DATE]
stock[is.na(portfolioWeight), portfolioWeight := 0]
stock[, benchmarkWeight := 1/.N, by=DATE]
fit <- fit.attribution(fitdata = stock, date.var = 'DATE', id.var = 'TICKER', return.var = 'RETURN',
                       weight.var = 'LOG.MARKETCAP', exposure.vars = c('NET.SALES','BOOK2MARKET','GICS.SECTOR'),
                       rob.stats = TRUE, full.resid.cov = FALSE, z.score = FALSE,
                       stdReturn = TRUE, resid.EWMA = TRUE)
cfit <-fit.contribution(fit, bm.wgt.var = 'benchmarkWeight', port.wgt.var = 'portfolioWeight')
}
\references{
David R Carino, Combining Attribution Effects Over Time, The Journal of Performance Measurement, pages 5 - 14.
}
\author{
Roger J. Bos, \email{roger.bos@gmail.com}
}
