% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.factor.models.R
\encoding{UTF-8}
\name{fit.risk.summary.report}
\alias{fit.risk.summary.report}
\title{fit.risk.summary.report}
\usage{
fit.risk.summary.report(riskmod, bm.wgt.var, port.wgt.var,
  filename = NULL)
}
\arguments{
\item{riskmod}{object returned by fit.fundamental() or fit.statistical()}

\item{bm.wgt.var}{benchmark weights, can be either of string of the name of the variable that contains the weights or a vector of the weights}

\item{port.wgt.var}{portfolio weights, can be either of string of the name of the variable that contains the weights or a vector of the weights}

\item{filename}{if not NULL, results will be saved to a CSV file with the given filename}
}
\value{
list of portfolioTbl and securityTbl.
The portfolio table contains the following data items for the portfolio:
*Portfolio.Holdings* - number of portfolio holdings
*Benchmark.Holdings* - number of benchmark holdings
*Total.Risk* - predicted standard deviation of the portfolio, aka absolute risk
*Benchmark.Risk* - predicted standard deviation of the benchmark
*Active.Risk* - predicted tracing error of the portfolio to the benchmark
*R.Squared* - the fraction of portfolio varinace explained by the benchmark
*Predicted.Beta* - forecasted beta based on the risk model
*Specific.Risk.Pct* - percent of *Total.Risk* not explaned by the risk model factors, aka idiosyncratic risk
*Factor.Risk.Pct* - percent of *Total.Risk* explaned by the risk model factors
The security table contains the following data items for each security:
*portfolio.wgt* - each stock's weight in the portfolio
*benchmark.wgt* - each stock's weight in the benchmark
*active.wgt* - each stock's active weight (portfolio minus benchmark)
*ivol* - each stock's idiosyncratic volatility, aka stock specific risk
*predicted.beta* - each stock's predicted beta
*risk.contribution* - each stock's contribution to *Total.Risk*
*factor.exposures* - each stock's exposure to each factor in the risk model
}
\description{
Function to create risk model report from fundamental or statistical risk model output
}
\examples{

### Statistical

load("fit.factor.models/data/Stock.df.Rdata")
retMat <- fit.data.cast(stock, item='RETURN', id.var = 'TICKER', date.var = 'DATE', reverse = TRUE)
fit <- fit.statistical(retMat)

stock <- as.data.table(stock)
stock[TICKER \%in\% c('SUNW','ORCL','MSFT'), portfolioWeight := 1/3, by=DATE]
stock[is.na(portfolioWeight), portfolioWeight := 0]
stock[, benchmarkWeight := 1/.N, by=DATE]
cols <- c("TICKER","portfolioWeight")
portfolioWeight <- stock[DATE == max(DATE), ..cols]
cols <- c("TICKER","benchmarkWeight")
benchmarkWeight <- stock[DATE == max(DATE), ..cols]

risk <- fit.risk.summary.report(riskmod = fit, bm.wgt.var = benchmarkWeight, port.wgt.var = portfolioWeight, filename = NULL)
risk$portfolioTbl
risk$securityTbl[Portfolio.Wgt > 0, ]


### Fundamental

load("fit.factor.models/data/Stock.df.Rdata")
stock <- as.data.table(stock)
stock[TICKER \%in\% c('SUNW','ORCL','MSFT'), portfolioWeight := 1/3, by=DATE]
stock[is.na(portfolioWeight), portfolioWeight := 0]
stock[, benchmarkWeight := 1/.N, by=DATE]
fit <- fit.fundamental(fitdata = stock, date.var = 'DATE', id.var = 'TICKER', return.var = 'RETURN',
                       weight.var = 'LOG.MARKETCAP', exposure.vars = c('NET.SALES','BOOK2MARKET','GICS.SECTOR'),
                       rob.stats = TRUE, z.score = FALSE, stdReturn = TRUE, calc.inv = TRUE, cov.wgt = FALSE, parkinson = FALSE)

stock <- as.data.table(stock)
stock[TICKER \%in\% c('SUNW','ORCL','MSFT'), portfolioWeight := 1/3, by=DATE]
stock[is.na(portfolioWeight), portfolioWeight := 0]
stock[, benchmarkWeight := 1/.N, by=DATE]
cols <- c("TICKER","portfolioWeight")
portfolioWeight <- stock[DATE == max(DATE), ..cols]
cols <- c("TICKER","benchmarkWeight")
benchmarkWeight <- stock[DATE == max(DATE), ..cols]

risk <- fit.risk.summary.report(riskmod = fit, bm.wgt.var = benchmarkWeight, port.wgt.var = portfolioWeight, filename = NULL)
risk$portfolioTbl
risk$securityTbl[Portfolio.Wgt > 0, ]

}
\author{
Roger J. Bos, \email{roger.bos@gmail.com}
}
\concept{factor models for attribution and risk}
