% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.factor.models.R
\docType{package}
\encoding{UTF-8}
\name{fit.statistical}
\alias{fit.statistical}
\alias{fit.statistical-package}
\title{fit.statistical}
\usage{
fit.statistical(retMat, use.cor = FALSE, erank = FALSE)
}
\arguments{
\item{retMat}{matrix (or data.table/data.frame) of returns.  Does not matter if the dates are in ascending or descending order.}

\item{use.cor}{boolean to use correlation matrix instead of covariance matrix, default is FALSE}

\item{erank}{boolean to use the effective rank instead of minimization algorithm, default is FALSE}
}
\value{
An object with S3 class "fir.factor.models" containing:

specific.risk is the vector idiosyncratic volatility (or specific risk),

factor.loadings is the matrix of loadings,

factor.cov is the matrix of factor covariances,

cov.mat is the matrix of variance/covariances,

inverse.cov is the inverse of the variance/covariance matrix,

princomp is the matrix of principal components (eigenvectors) of the fitted risk model,

dates is the vector of the dates used to estimate the risk model, and

id is the vector is ids representing the securities in the estimation universe.
}
\description{
Function to calculate a statistical risk model (using principal component analysis). The code has been adapted from Kakushadze and Yu (2017).
}
\examples{
retMat <- fit.data.cast(stock, item='RETURN', id.var = 'TICKER', date.var = 'DATE', reverse = TRUE)
fit <- fit.statistical(retMat)
names(fit)
}
\references{
Kakushadze, Zura and Yu, Willie, Statistical Risk Models (February 14, 2016). The Journal of Investment Strategies 6(2) (2017) 1-40. Available at SSRN: \url{https://ssrn.com/abstract=2732453}
}
\author{
Roger J. Bos, \email{roger.bos@gmail.com}
}
\concept{factor models for attribution and risk}
