% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.factor.models.R
\encoding{UTF-8}
\name{fit.zscore}
\alias{fit.zscore}
\title{fit.zscore}
\usage{
fit.zscore(x, dte, w = 1, grp = 1, rob.stats = FALSE)
}
\arguments{
\item{x}{vector of data to be standardized}

\item{dte}{vector of dates (or anything else) used to partition the data}

\item{w}{optional vector of weights used to calculate weighted mean, defaults to equal weighted}

\item{grp}{optional vector of group designations used to standardize within groups, defaults to a signle group}

\item{rob.stats}{logical parameter to subtract the the median rather than the mean, defaults to FALSE}
}
\value{
vector of standardized data with a mean near zero and a standard deviation near one
}
\description{
Function to perform cross-sectional standardization of raw data (vectorized)
}
\examples{
fit.zscore(x = data[['mcap']), dte = data[[date]]), w = data[['weight']]), rob.stats = TRUE)
}
\author{
Roger J. Bos, \email{roger.bos@gmail.com}
}
\concept{factor models for attribution and risk}
